% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Mutation_functions.R
\name{add_deletion}
\alias{add_deletion}
\title{Function to add deletion to gene map (chromosomal location data frame)}
\usage{
add_deletion(gm, Ref_start, Ref_end, Chr)
}
\arguments{
\item{gm}{Chromosomal location data frame}

\item{Ref_start}{Starting position of deletion}

\item{Ref_end}{Final position of deletion}

\item{Chr}{Chromosome name}
}
\value{
Chromosomal location data frame with additional deletion info
}
\description{
Function to add deletion to gene map (chromosomal location data frame)
}
\examples{
gene_map  =  tugHall_dataset$gene_map
gene_map$pnts = ''
gm_new = add_deletion( gm = gene_map, Ref_start = 112775658, Ref_end = 112775716, Chr = '5')
# to check add_del ... add_dupl, please, use diff library and:
# diff_data( gm_ref, gm, show_unchanged_columns = TRUE, always_show_order = TRUE )
}
