% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Mutation_functions.R
\name{add_pnt_mutation}
\alias{add_pnt_mutation}
\title{Function to add point mutation to data.frame gene_map (chromosomal location data frame)}
\usage{
add_pnt_mutation(gm = gm, pos_pnt, Chr = Chr)
}
\arguments{
\item{gm}{Chromosomal location data frame}

\item{pos_pnt}{Position of point mutation}

\item{Chr}{Chromosome name}
}
\value{
Chromosomal location data frame with additional point mutation info
}
\description{
Function to add point mutation to data.frame gene_map (chromosomal location data frame)
}
\examples{
gene_map = tugHall_dataset$gene_map
gene_map$pnts  =  ''
gm2 = add_pnt_mutation( gm = gene_map, pos_pnt = 112775637 , Chr = '5' )
}
