% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Model_functions.R
\name{calc_binom}
\alias{calc_binom}
\title{Function to calculate binomial distribution including BIG NUMBERS like 10^12 and more using approximation with normal distribution}
\usage{
calc_binom(tr, n, p)
}
\arguments{
\item{tr}{Length of vector with successes trials}

\item{n}{Number of independent Bernoulli trials}

\item{p}{Probability to get successes in trials}
}
\value{
Vector of integer numbers of successes trials
}
\description{
Function to calculate binomial distribution including BIG NUMBERS like 10^12 and more using approximation with normal distribution
}
\examples{
calc_binom(tr = 3, n = 40, p = 0.9)
calc_binom(tr = 3, n = 4E20, p = 9E-9)
}
