% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Model_functions.R
\name{change_pnt_by_cna}
\alias{change_pnt_by_cna}
\title{Function to change the point mutation due to CNA}
\usage{
change_pnt_by_cna(pnt1, start_end, t)
}
\arguments{
\item{pnt1}{Object of class 'Point_Mutations'}

\item{start_end}{Vector with initial and final positions of CNA}

\item{t}{'dup' or 'del' for duplication or deletion respectively}
}
\value{
NULL, but pnt1 data is updated due to CNA
}
\description{
Function to change the point mutation due to CNA
}
\examples{
pnt1 = tugHall_dataset$pnt_clones[[ 1 ]]
start_end = c( pnt1$Phys_pos - 50 , pnt1$Phys_pos + 50  )
message( pnt1$Copy_number )
change_pnt_by_cna( pnt1, start_end, t = 'dup' ) # View( pnt1 )
message( pnt1$Copy_number )
change_pnt_by_cna( pnt1, start_end, t = 'del' ) # View( pnt1 )
message( pnt1$Copy_number )
}
