% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_pkg}
\alias{check_pkg}
\title{Check the installation of a package for some functions}
\usage{
check_pkg(pkg)
}
\arguments{
\item{pkg}{Package name}
}
\value{
if the package is installed then it returns NULL else it returns error message
}
\description{
Check the installation of a package for some functions
}
\examples{
check_pkg( pkg = 'grDevices' )
}
