% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Mutation_functions.R
\name{check_pnts}
\alias{check_pnts}
\title{Function to check what pnts do fall into the range?}
\usage{
check_pnts(gm_w1)
}
\arguments{
\item{gm_w1}{A row from data.frame gene_map}
}
\value{
Return the point mutations which fall into the range
}
\description{
Function to check what pnts do fall into the range?
}
\examples{
gene_map = tugHall_dataset$gene_map
gene_map$pnts = ''
gene_map$pnts[6] = '112792451, 112792442'
gm_w1 = gene_map[6,]
check_pnts( gm_w1 )
}
