% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Input.R
\name{check_previous_data}
\alias{check_previous_data}
\title{Function to check the files from the previous simulation}
\usage{
check_previous_data()
}
\value{
\code{check_previous_data} returns NULL and renames Output folder as well as monitoring file
to the folder and file with time stamp
}
\description{
Function to check the files from the previous simulation are exist and if so to move
all of them to the folder with name \code{ /Output[Time.stamp]/ },
the \code{ [Time.stamp]/ } in the format \code{2022_10_22_15_51_09} or \code{year_month_day_hour_min_sec}
}
\examples{
NULL
}
