% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Classes.R
\docType{class}
\name{Clone-class}
\alias{Clone-class}
\alias{clone}
\title{Class 'Clone' for clones}
\description{
Class 'Clone' for clones
}
\section{Fields}{

\describe{
\item{\code{id}}{numeric. ID of a clone}

\item{\code{parent}}{numeric. Parent ID (for first - 0)}

\item{\code{N_cells}}{numeric. Number of cells in clone}

\item{\code{c}}{numeric. Split counter as average value for all cells in clone}

\item{\code{d}}{numeric. Probability of division}

\item{\code{i}}{numeric. Probability of Hayflick limit}

\item{\code{m}}{numeric. Probability that gene normal function is destroyed due to epigenome abnormality / mutation rate}

\item{\code{a}}{numeric. Probability of apoptosis for a cell in the clone}

\item{\code{s}}{numeric. Coefficient in the sigmoid function of the mutation density}

\item{\code{k}}{numeric. Probability of cell death by environment}

\item{\code{E}}{numeric. Coefficient of friction term against to the split probability.}

\item{\code{Nmax}}{numeric. Coefficient for determination the max number of cells \cr
that can exist in the primary tumor (Nmax = 1/E) \cr}

\item{\code{im}}{numeric. Probability of the invasion/ metastatic transformation}

\item{\code{Ha}}{numeric. Apoptosis hallmark value}

\item{\code{Him}}{numeric. Invasion/ metastasis hallmark}

\item{\code{Hi}}{numeric. Mitotic restriction hallmark (immortalization hallmark)}

\item{\code{Hd}}{numeric. Growth/antigrowth hallmark (division rate hallmark)}

\item{\code{Hb}}{numeric. Angiogenesis hallmark}

\item{\code{gene}}{numeric. Vector of flags for each genes if they have driver mutation}

\item{\code{pasgene}}{numeric. Vector of flags for each genes if they have passenger mutation}

\item{\code{PointMut_ID}}{numeric. ID of point mutation in list of objects of class 'Point_Mutations'}

\item{\code{CNA_ID}}{numeric. ID of CNA mutation in list of objects of class 'CNA_Mutations'}

\item{\code{mutden}}{numeric. Gene mutation density}

\item{\code{invasion}}{logical. Indicator that clone is metastatic (invasion/metastatic transformation occured or not)}

\item{\code{primary}}{logical. Logical variable is clone primary tumor or not (normal)}

\item{\code{birthday}}{numeric. Time step of birth of clone}
}}


\examples{
clone = tugHall_dataset$clones[[ 1 ]]
print(clone$Ha)
print(clone$N_cells)
clone$calcApoptosis()    # to calculate apoptosis death probability based on mutation density
}
