% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Classes.R
\docType{class}
\name{CNA_Mutations-class}
\alias{CNA_Mutations-class}
\alias{CNA_Mutations}
\title{Class 'CNA_Mutations'}
\description{
Class 'CNA_Mutations'
}
\section{Fields}{

\describe{
\item{\code{CNA_ID}}{numeric. ID of CNA mutation}

\item{\code{Parental_1or2}}{numeric. Parental chromosome, could be 1 or 2}

\item{\code{dupOrdel}}{character. dup for duplication or del for deletion}

\item{\code{Chr}}{character. Chromosome name}

\item{\code{Ref_start}}{numeric. Reference start position}

\item{\code{Ref_end}}{numeric. Reference final position}

\item{\code{Gene_names}}{character. Names of genes involved in CNA}

\item{\code{MalfunctionedByCNA}}{logical.}

\item{\code{mut_order}}{numeric. True for driver mutation and False for passenger mutation}
}}


\examples{
cna = tugHall_dataset$cna_clones[[ 1 ]]
cna$safe()   # to save as row of data.frame
cna$copy()
cna$initialize()
cna$show()   # After initialization
}
