% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Mutation_functions.R
\name{copy_CNA}
\alias{copy_CNA}
\title{Function to copy CNA info}
\usage{
copy_CNA(CNA1)
}
\arguments{
\item{CNA1}{Object of class 'CNA_Mutations'}
}
\value{
The same object of class 'CNA_Mutations'
}
\description{
Function to copy CNA info
}
\examples{
cna = tugHall_dataset$cna_clones[[ 1 ]]
cna2 = copy_CNA( cna )
cna$safe()
cna2$safe()
}
