% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_files_to_Input}
\alias{copy_files_to_Input}
\title{Function to copy the files by default from extdata folder in the library to Input folder in the working directory}
\usage{
copy_files_to_Input(
  files = c("CCDS.current.txt", "CF.txt", "cloneinit.txt", "gene_hallmarks.txt",
    "gene_map.txt", "parameters.txt"),
  dir = "Input"
)
}
\arguments{
\item{files}{Files to copy, vector of names of files by default: \cr
files = c( 'CCDS.current.txt', 'CF.txt', 'cloneinit.txt', 'gene_hallmarks.txt','gene_map.txt','parameters.txt' )}

\item{dir}{Folder to where files should be save, by default dir = 'Input'}
}
\value{
List of logic numbers for each copied file, TRUE - success, FALSE - not success
}
\description{
Function to copy the files by default from extdata folder in the library to Input folder in the working directory
}
\examples{
files = c('CF.txt', 'cloneinit.txt','gene_hallmarks.txt','gene_map.txt','parameters.txt' )
copy_files_to_Input( files, dir = 'Input' )
}
