% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_files_to_Output}
\alias{copy_files_to_Output}
\title{Function to copy the files of an example of simulation or from '/extdata/Output/' folder in the library to '/Output/' folder in the working directory}
\usage{
copy_files_to_Output(
  files = c("cloneout.txt", "CNA_mutations.txt", "point_mutations.txt", "gene_MAP.txt",
    "geneout.txt", "log.txt", "order_genes_dysfunction.txt", "VAF_data.txt", "VAF.txt",
    "weights.txt"),
  dir = "Output"
)
}
\arguments{
\item{files}{Files to copy, vector of names of files by default: \cr
files = c( 'cloneout.txt', 'CNA_mutations.txt', 'point_mutations.txt', 'gene_MAP.txt','geneout.txt','log.txt', 'order_genes_dysfunction.txt', 'VAF_data.txt', 'VAF.txt', 'weights.txt' )}

\item{dir}{Folder to where files should be save, by default dir = 'Output'}
}
\value{
List of logic numbers for each copied file, TRUE - success, FALSE - not success
}
\description{
Function to copy the files of an example of simulation or from '/extdata/Output/' folder in the library to '/Output/' folder in the working directory
}
\examples{
files = c( 'cloneout.txt', 'CNA_mutations.txt', 'point_mutations.txt', 'gene_MAP.txt')
copy_files_to_Output( files )
files = c('geneout.txt','log.txt', 'VAF_data.txt', 'VAF.txt', 'weights.txt' )
copy_files_to_Output( files )
}
