% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Mutation_functions.R
\name{copy_pnt}
\alias{copy_pnt}
\title{Function to copy of point mutation info}
\usage{
copy_pnt(pnt1)
}
\arguments{
\item{pnt1}{Object of class 'Point_Mutations'}
}
\value{
The same object of class 'Point_Mutations' with the same ID
}
\description{
Function to copy of point mutation info
}
\examples{
pnt = tugHall_dataset$pnt_clones[[ 1 ]]
copy_pnt( pnt ) # View( pnt )
}
