% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Mutation_functions.R
\name{copy_pnt_no_mutation}
\alias{copy_pnt_no_mutation}
\title{Function to copy of pnt1 without mutation info for allele A}
\usage{
copy_pnt_no_mutation(pnt1)
}
\arguments{
\item{pnt1}{Object of class 'Point_Mutations'}
}
\value{
Object of class 'Point_Mutations' for another chromosome
}
\description{
Function to copy of pnt1 without mutation info for allele A
}
\examples{
pnt = tugHall_dataset$pnt_clones[[ 1 ]]
copy_pnt_no_mutation( pnt )  # View( pnt )
}
