% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Input.R
\name{define_compaction_factor}
\alias{define_compaction_factor}
\title{Define compaction factor}
\usage{
define_compaction_factor(
  cf = data.frame(Ha = 1, Hb = 1, Hd = 1, Hi = 1, Him = 1),
  read_fl = TRUE,
  file_name = "./Input/CF.txt"
)
}
\arguments{
\item{cf}{Data frame with compaction factors for all the hallmarks, for example, data.frame( Ha = 1, Hb = 1, Hd = 1, Hi = 1, Him = 1 )}

\item{read_fl}{Indicator to read file or not, logical type only}

\item{file_name}{File name to rad all the parameters, it is used only if read_fl == TRUE}
}
\value{
Data frame with with compaction factors for all the hallmarks
}
\description{
Define compaction factor
}
\examples{
copy_files_to_Input()
define_compaction_factor( read_fl = TRUE , file_name = './Input/CF.txt' )
CF1 = CF
cf = data.frame( Ha = 0.1, Hb = 0.2, Hd = 0.7, Hi = 1, Him = 0.5 )
define_compaction_factor( cf = cf, read_fl = FALSE )  # View( c( CF, CF1 ) ) to compare
}
