% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Input.R
\name{define_files_names}
\alias{define_files_names}
\title{Function to define all the files names}
\usage{
define_files_names(
  mainDir = getwd(),
  sbdr_Input = "/Input",
  sbdr_Output = "/Output"
)
}
\arguments{
\item{mainDir}{Working directory}

\item{sbdr_Input}{Sub directory for input files}

\item{sbdr_Output}{Sub directory for output files}
}
\value{
NULL, but all file names are defined in GLOBAL environment
}
\description{
Function to define all the files names
}
\examples{
define_files_names()
}
