% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Input.R
\name{define_gene_location}
\alias{define_gene_location}
\title{Define genes' location in chromosome}
\usage{
define_gene_location(
  file_input = "Input/CCDS.current.txt",
  genes_list = c("CCDS4107.1", "CCDS8702.1", "CCDS43171.1", "CCDS11118.1")
)
}
\arguments{
\item{file_input}{is a name of file to input where the information about genes location is defined. That is loaded from CCDS database https://ftp.ncbi.nlm.nih.gov/pub/CCDS/current_human/}

\item{genes_list}{is a list of genes' names like CCDS4107.1 in the CCDS database.}
}
\value{
Function returns the table of genes' locations in DNA
}
\description{
Define genes' location in chromosome
}
\examples{
copy_files_to_Input()
define_gene_location()
file_input  =  'Input/CCDS.current.txt'
genes_list  =  c( 'CCDS4107.1', 'CCDS8702.1', 'CCDS43171.1', 'CCDS11118.1' )
define_gene_location( file_input = file_input,  genes_list = genes_list )
}
