% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{define_par_for_plot}
\alias{define_par_for_plot}
\title{Function to change par for plots and after plotting it returns par values}
\usage{
define_par_for_plot(change_par_back)
}
\arguments{
\item{change_par_back}{logical. If TRUE it changes par options back after finishing a function call}
}
\value{
Change par() options and returns it (after finishing the function) to values before a function has started
}
\description{
Function to change par for plots and after plotting it returns par values
}
\examples{
define_par_for_plot( change_par_back = TRUE )
}
