% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Input.R
\name{define_parameters}
\alias{define_parameters}
\title{Define all the parameters for a simulation}
\usage{
define_parameters(
  E0 = 1e-04,
  F0 = 10,
  m0 = 1e-07,
  uo = 0.9,
  us = 0.9,
  s0 = 10,
  k0 = 0.12,
  d0 = 0.4,
  censore_n = 10^5,
  censore_t = 50,
  m_dup = 1e-08,
  m_del = 1e-08,
  lambda_dup = 5000,
  lambda_del = 7000,
  uo_dup = 0.8,
  us_dup = 0.5,
  uo_del = 0,
  us_del = 0.8,
  Compaction_factor = TRUE,
  model = c("proportional_metastatic", "threshold_metastatic", "simplified")[1],
  time_stop = 120,
  read_fl = FALSE,
  file_name = "./Input/parameters.txt",
  n_repeat = 1000,
  monitor = TRUE
)
}
\arguments{
\item{E0}{Parameter in the division probability, numeric type only}

\item{F0}{Parameter in the division probability, numeric type only}

\item{m0}{Mutation probability for point mutation, numeric type only}

\item{uo}{Oncogene mutation probability, numeric type only}

\item{us}{Suppressor mutation probability, numeric type only}

\item{s0}{Parameter in the sigmoid function, numeric type only}

\item{k0}{Environmental death probability, numeric type only}

\item{d0}{Initial probability to divide cells, numeric type only}

\item{censore_n}{Max cell number where the program forcibly stops, integer type only}

\item{censore_t}{Max time where the program forcibly stops, integer type only}

\item{m_dup}{Mutation probability for duplication, numeric type only}

\item{m_del}{Mutation probability for deletion, numeric type only}

\item{lambda_dup}{CNA duplication average length (of the geometrical distribution for the length), integer type only}

\item{lambda_del}{CNA deletion average length (of the geometrical distribution for the length), integer type only}

\item{uo_dup}{Gene malfunction probability by CNA duplication for oncogene, numeric type only}

\item{us_dup}{Gene malfunction probability by CNA duplication for suppressor, numeric type only}

\item{uo_del}{Gene malfunction probability by CNA deletion    for oncogene, numeric type only}

\item{us_del}{Gene malfunction probability by CNA deletion    for suppressor, numeric type only}

\item{Compaction_factor}{Compaction factor, logical indicator. True means 'to use', False means 'do not use' Compaction factor for hallmarks variables}

\item{model}{Name of the model to use. Can be  'proportional_metastatic' or 'threshold_metastatic' or 'simplified'}

\item{time_stop}{Max time in seconds of running after that the program forcibly stops, integer type only}

\item{read_fl}{Indicator to read file or not, logical type only}

\item{file_name}{File name to rad all the parameters, it is used only if read_fl == TRUE}

\item{n_repeat}{Max number of repetition of the program until the NON-ZERO output will be getting, integer type only}

\item{monitor}{The indicator to make monitor file during a simulation or do not make, logical type only}
}
\value{
Values of all the parameters
}
\description{
Define all the parameters for a simulation
}
\examples{
copy_files_to_Input()
define_parameters( read_fl = TRUE , file_name = './Input/parameters.txt' )
define_parameters( read_fl = FALSE )
}
