% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_intervention.R
\name{drug_intervention}
\alias{drug_intervention}
\title{Function to emulate drug intervention to pool of tumor cells by killing a part of these cells
and blocking a part of clones corresponding to malfunctioned gene}
\usage{
drug_intervention(kill_prob = 0, block_prob = 1, gene)
}
\arguments{
\item{kill_prob}{Probability of killing cancer cells corresponding to the malfunctioned gene}

\item{block_prob}{Probability of blocking cancer cells corresponding to the malfunctioned gene}

\item{gene}{Name of target gene to kill and block tumor cells by a drug}
}
\value{
NULL changing clones and onco_clones objects in tugHall environment pck.env
}
\description{
Function to emulate drug intervention to pool of tumor cells by killing a part of these cells
and blocking a part of clones corresponding to malfunctioned gene
}
