% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Classes.R
\docType{class}
\name{Environ-class}
\alias{Environ-class}
\alias{environ}
\title{Class 'Environ'}
\description{
Class 'Environ'
}
\section{Fields}{

\describe{
\item{\code{T}}{numeric. Time counter}

\item{\code{N}}{numeric. Number of normal cells}

\item{\code{P}}{numeric. Number of primary tumor cells}

\item{\code{M}}{numeric. Number of metastatic cells}

\item{\code{F}}{numeric. Coefficient that determines the maximal number of cells in pool of primary tumor cells}

\item{\code{c}}{numeric. Average number of divisions in pool of clones}

\item{\code{d}}{numeric. Mean value of splitting probability}

\item{\code{i}}{numeric. Average value of immortalization probability}

\item{\code{a}}{numeric. Average value of apoptosis probability}

\item{\code{k}}{numeric. Average probability of cell death via environment death}

\item{\code{E}}{numeric. Average value of coefficients of friction term}

\item{\code{Nmax}}{numeric. Maximal number of primary tumor cells that can exist in pool of clones}

\item{\code{im}}{numeric. Average value of invasion/metastasis probability}

\item{\code{Ha}}{numeric. Average value of apoptosis hallmark Ha}

\item{\code{Him}}{numeric. Average value of invasion/metastasis hallmark Him}

\item{\code{Hi}}{numeric. Average value of immortalization hallmark Hi}

\item{\code{Hd}}{numeric. Average value of growth/antigrowth hallmark Hd}

\item{\code{Hb}}{numeric. Average value of angiogenesis hallmark Hb}

\item{\code{type}}{numeric. Invasion / metastatic ratio}

\item{\code{gene}}{numeric. Cancer gene damage rate}

\item{\code{mutden}}{numeric. Average mutation rate}

\item{\code{last_id}}{numeric. Maximal ID in the pool of clones.}
}}


\examples{
env = tugHall_dataset$env
print( env )
env$initFields()
}
