% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{foolproof}
\alias{foolproof}
\title{Foolproof function allows to checking the consistency of all the input parameters}
\usage{
foolproof()
}
\description{
Foolproof function allows to checking the consistency of all the input parameters.
It should be used just before a loop of simulation. So, all the parameters should be defined,
objects onco and hall should be initialized as well. The function checks the list of parameters' names,
absence of NA and NULL in the input data, self-consistency of genes names,
correctness of hallmarks values, and finally, that all the necessary information is defined.
}
\examples{
NULL
}
