% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_colors}
\alias{gen_colors}
\title{Function to make a large number of colors}
\usage{
gen_colors(nm = 12)
}
\arguments{
\item{nm}{Number of colors}
}
\value{
Vector of colors with length more than nm
}
\description{
Function to make a large number of colors
}
\examples{
clrs = gen_colors( nm = 120 )
}
