% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Mutation_functions.R
\name{generate_cna}
\alias{generate_cna}
\title{Function to generate object of CNA mutation}
\usage{
generate_cna(prntl, genes, start_end, onco1, dupOrdel)
}
\arguments{
\item{prntl}{The 1st or 2nd parental chromosome}

\item{genes}{Genes names}

\item{start_end}{vector with start and final positions of CNA}

\item{onco1}{Object of class 'OncoGene'}

\item{dupOrdel}{It could be 'dup' or 'del' to denote duplication or deletion}
}
\value{
Object of class 'CNA_Mutations'
}
\description{
Function to generate object of CNA mutation
}
\examples{
copy_files_to_Input()
load_tugHall.Environment( tugHall_dataset )
onco = tugHall_dataset$onco
gene_map = tugHall_dataset$gene_map
pnt_clones = tugHall_dataset$pnt_clones
cna_clones = tugHall_dataset$cna_clones
mut_order = 234
start_end = c(112775658, 112775716 )
withr::with_environment( env = pck.env, code = generate_cna( prntl = 1, genes = 'APC', start_end = start_end, onco1 = onco, dupOrdel = 'dup' ) )
}
