% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Mutation_functions.R
\name{generate_pnt}
\alias{generate_pnt}
\title{Function to generate an object of class 'Point_Mutations'}
\usage{
generate_pnt(prntl, gene, pos, onco1, Chr, mutation = NA)
}
\arguments{
\item{prntl}{Parental chromosome, could be 1 or 2}

\item{gene}{Gene name}

\item{pos}{Position of point mutation}

\item{onco1}{Object of class 'OncoGene'}

\item{Chr}{Chromosome name}

\item{mutation}{If mutation is NOT NA then MalfunctionedByPointMut = TRUE, else it is defined by corresponding probabilities}
}
\value{
Object of class 'Point_Mutations'
}
\description{
Function to generate an object of class 'Point_Mutations'
}
\examples{
gm = tugHall_dataset$gene_map
gm_1_2 = list( gm, gm )
onco = tugHall_dataset$onco
pnt_clones = tugHall_dataset$pnt_clones
copy_files_to_Input()
define_parameters()
mut_order = 234  # As an example
pnt1 = generate_pnt( prntl = 1, gene = 'APC', pos = 112767192, onco, Chr = '5', mutation = NA )
}
