% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Mutation_functions.R
\name{generate_to_copy_pnt}
\alias{generate_to_copy_pnt}
\title{Function to generate the same object of class 'Point_Mutations' with coping all information from input object}
\usage{
generate_to_copy_pnt(pnt)
}
\arguments{
\item{pnt}{Object of class 'Point_Mutations'}
}
\value{
The same object of class 'Point_Mutations' with different ID
}
\description{
Function to generate the same object of class 'Point_Mutations' with coping all information from input object
}
\examples{
pnt = tugHall_dataset$pnt_clones[[ 1 ]]
pnt_clones = tugHall_dataset$pnt_clones
pnt2 = generate_to_copy_pnt( pnt )
}
