% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Mutation_functions.R
\name{get_cna_mutation}
\alias{get_cna_mutation}
\title{Generation CNA mutation info}
\usage{
get_cna_mutation(onco1, dupOrdel, gm_1_2)
}
\arguments{
\item{onco1}{Object of class 'OncoGene'}

\item{dupOrdel}{It could be 'dup' or 'del' to denote duplication or deletion}

\item{gm_1_2}{List of two data frames (for 1st and 2nd parental chromosomes) with genes' location information}
}
\value{
List of ( prntl - 1 or 2 parental chromosome, Chr - name of chromosome, genes - genes names, start_end - vector with start and end positions of CNA, w_cna - rows of CNA in gene_map data frame )
}
\description{
Generation CNA mutation info
}
\examples{
copy_files_to_Input()
load_tugHall.Environment( tugHall_dataset )
onco = tugHall_dataset$onco
gm = tugHall_dataset$gene_map
withr::with_environment( env = pck.env, code = get_cna_mutation( onco1 = onco, dupOrdel = 'dup', gm_1_2 = list(gm, gm) ) )
withr::with_environment( env = pck.env, code = get_cna_mutation( onco1 = onco, dupOrdel = 'del', gm_1_2 = list(gm, gm) ) )
}
