% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_clones.R
\name{get_flow_data}
\alias{get_flow_data}
\title{Function to get data about last simulation from cloneoutfile}
\usage{
get_flow_data(
  cloneoutfile,
  genefile,
  mainDir = getwd(),
  sbdr_Output = "/Output"
)
}
\arguments{
\item{cloneoutfile}{Name of file to read data about clone evolution}

\item{genefile}{Name of file with hallmarks values}

\item{mainDir}{Working directory, by default mainDir = getwd()}

\item{sbdr_Output}{Directory for output data getting from mainDir}
}
\value{
list of data.frames like onco, hall, data_last (data of last time step), \cr
data_avg (average data for all time steps), data_flow (data without average rows), time_max (max time step), \cr
pnt_mut and pnt_mut_B (data.frame of point mutations for both alleles and for allele B only ) \cr
and cna_mut (data.frame of CNA mutations)
}
\description{
Function to get data about last simulation from cloneoutfile
}
\examples{
copy_files_to_Input()
define_files_names()
define_parameters( read_fl = TRUE , file_name = './Input/parameters.txt' )
define_compaction_factor( read_fl = TRUE , file_name = './Input/CF.txt' )
define_gene_location()
copy_files_to_Output()
dataset = get_flow_data(cloneoutfile, genefile, mainDir = getwd(), sbdr_Output = '/Output' )
# View(dataset)
}
