% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_maps.R
\name{get_len_cds_rna}
\alias{get_len_cds_rna}
\title{Function to get length of CDS and whole gene from gene_map data.frame}
\usage{
get_len_cds_rna(gene_map)
}
\arguments{
\item{gene_map}{data.frame with info about genes' locations}
}
\value{
list of ( Name, CDS, LEN_Genes ) where Name is a vector of genes' names, CDS is a vector of CDS lengths, LEN_Genes is a vector of length of whole genes including introns and exons
}
\description{
Function to get length of CDS and whole gene from gene_map data.frame
}
\examples{
gene_map = tugHall_dataset$gene_map
onco = tugHall_dataset$onco
get_len_cds_rna( gene_map)
}
