% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Mutation_functions.R
\name{get_point_mutation_for_gene}
\alias{get_point_mutation_for_gene}
\title{Generation point mutation info for the particular gene}
\usage{
get_point_mutation_for_gene(onco1, gm_1_2, gene)
}
\arguments{
\item{onco1}{Object of class 'OncoGene'}

\item{gm_1_2}{List of two data frames (for 1st and 2nd parental chromosomes) with genes' location information}

\item{gene}{Gene's name where point mutation should be occured}
}
\value{
list of (prntl - 1 or 2 parental chromosome, gene - gene name, pos - position of point mutation, Chr - name of chromosome )
}
\description{
Generation point mutation info for the particular gene
}
\examples{
gm = tugHall_dataset$gene_map
gm_1_2 = list( gm, gm )
onco = tugHall_dataset$onco
get_point_mutation_for_gene( onco, gm_1_2, gene = 'APC')
get_point_mutation_for_gene( onco, gm_1_2, gene = 'KRAS')
}
