% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_clones.R
\name{get_rho_VAF}
\alias{get_rho_VAF}
\title{Function to get Variant allele frequencies (VAF) based on rho input parameters}
\usage{
get_rho_VAF(vf = NULL, rho = c(0, 0.1, 0.5), file_name = "./Output/VAF.txt")
}
\arguments{
\item{vf}{data.frame getting from get_VAF() function}

\item{rho}{Vector of rho parameter in the range (0,1)}

\item{file_name}{Name of file to save VAF}
}
\value{
VAF for different rho with separation for metastatic cells and (primary tumor + speckled normal) cells
}
\description{
Function to get Variant allele frequencies (VAF) based on rho input parameters
}
\examples{
pnt_mut  =  tugHall_dataset$pnt_mut
data_last  =  tugHall_dataset$data_last
if ( !dir.exists('./Output') ) dir.create('./Output')
vf = get_VAF( pnt_mut, data_last, file_name = 'Output/VAF_data.txt')
VAF = get_rho_VAF( vf = vf, rho = c( 0.0, 0.1, 0.5 ) , file_name = './Output/VAF.txt' )
}
