% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Model_functions.R
\name{get_type}
\alias{get_type}
\title{Function to get type of the clone: normal, primary or metastatic}
\usage{
get_type(clone1)
}
\arguments{
\item{clone1}{Object of class 'Clone'}
}
\value{
One of characters 'normal', 'primary' or 'metastatic'
}
\description{
Function to get type of the clone: normal, primary or metastatic
}
\examples{
clone1 = tugHall_dataset$clones[[1]]
get_type( clone1 )
clone1 = tugHall_dataset$clones[[56]]
get_type( clone1 )
}
