% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_clones.R
\name{get_VAF}
\alias{get_VAF}
\title{Function to get data about Variant allele frequencies (VAF)}
\usage{
get_VAF(pnt_mut, data_last, file_name = "Output/VAF_data.txt")
}
\arguments{
\item{pnt_mut}{data.frame with point mutation info}

\item{data_last}{data.frame with data of simulation at the last time step}

\item{file_name}{Name of file to save data}
}
\value{
data.frame with info about Variant allele frequencies
}
\description{
Function to get data about Variant allele frequencies (VAF)
}
\examples{
pnt_mut  =  tugHall_dataset$pnt_mut
data_last  =  tugHall_dataset$data_last
vf = get_VAF( pnt_mut, data_last, file_name = 'Output/VAF_data.txt')
}
