% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Classes.R
\docType{class}
\name{HallMark-class}
\alias{HallMark-class}
\alias{hallmark}
\title{Class 'HallMark'}
\description{
Class 'HallMark'
}
\section{Fields}{

\describe{
\item{\code{Ha}}{numeric. Apoptosis hallmark indexes of genes in onco$name, \cr where onco is object of class OncoGene}

\item{\code{Hi}}{numeric. Immortalization hallmark indexes of genes in onco$name, \cr where onco is object of class OncoGene}

\item{\code{Hd}}{numeric. Growth/antigrowth hallmark indexes of genes in onco$name, \cr where onco is object of class OncoGene}

\item{\code{Hb}}{numeric. Angiogenesis hallmark indexes of genes in onco$name, \cr where onco is object of class OncoGene}

\item{\code{Him}}{numeric. Invasion/metastatic transformation hallmark indexes of genes in onco$name, \cr where onco is object of class OncoGene}

\item{\code{Ha_w}}{numeric. Apoptosis hallmark weights of genes}

\item{\code{Hi_w}}{numeric. Immortalization hallmark weights of genes}

\item{\code{Hd_w}}{numeric. Growth/antigrowth hallmark weights of genes}

\item{\code{Hb_w}}{numeric. Angiogenesis hallmark weights of genes}

\item{\code{Him_w}}{numeric. Invasion/metastatic transformation hallmark weights of genes}

\item{\code{notHa}}{numeric. Indexes of genes which are not in apoptosis hallmark}
}}


\examples{
hall = tugHall_dataset$hall
print( hall )
hall$copy()
hall$show()
}
