% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Model_functions.R
\name{init_clones}
\alias{init_clones}
\title{Function to read file with initial clones}
\usage{
init_clones(clonefile, clone1)
}
\arguments{
\item{clonefile}{File to read}

\item{clone1}{Object of class 'Clone'}
}
\value{
List of objects of class 'Clone
}
\description{
Function to read file with initial clones
}
\examples{
copy_files_to_Input()
define_files_names()
env = tugHall_dataset$env
define_parameters()
onco = tugHall_dataset$onco
clone1 = tugHall_dataset$clones[[ 1 ]]
clones = init_clones(clonefile, clone1)  # View( clones )
}
