% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Model_functions.R
\name{init_pnt_clones}
\alias{init_pnt_clones}
\title{Function to generate point mutations for initial clones}
\usage{
init_pnt_clones(clones, onco_clones)
}
\arguments{
\item{clones}{List of objects of class 'Clone'}

\item{onco_clones}{List of objects of class 'OncoGene'}
}
\description{
Function to generate point mutations for initial clones
}
\examples{
clones = tugHall_dataset$clones
define_parameters()
onco = tugHall_dataset$onco
onco_clones = tugHall_dataset$onco_clones
copy_files_to_Input()
copy_files_to_Output()
define_gene_location()
pnt_clones = tugHall_dataset$pnt_clones
cna_clones = tugHall_dataset$cna_clones
mut_order = 234
\dontrun{
init_pnt_clones( clones, onco_clones )  # change pnt_clones for initialization
}
}
