% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{make_input_format}
\alias{make_input_format}
\alias{make_input_range}
\alias{make_input_dataset}
\title{Function to prepare dataset of input parameters for parallel calculations}
\usage{
make_input_format(
  par_exclude = c("censor_cells_number", "censor_time_step", "clonefile", "cloneoutfile",
    "ctmax", "genefile", "geneoutfile", "lambda_del", "lambda_dup", "logoutfile",
    "model_name", "monitor", "n_repeat", "real_time_stop",
    "tumbler_for_metastasis_trial", "tumbler_for_apoptosis_trial",
    "tumbler_for_immortalization_trial", "tumbler_for_angiogenesis_trial",
    "tumbler_for_drug_intervention_trial")
)

make_input_range(frmt)

make_input_dataset(
  frmt,
  rng,
  n_simulations = 10,
  discrete = TRUE,
  n_graduations = 11
)
}
\arguments{
\item{par_exclude}{List of parameters to exclude from data frame of
input parameters because they will be constant for all the simulations}

\item{frmt}{List of results of function \code{make_input_format()} as input format for the range of each parameter}

\item{rng}{Data frame was gotten as a result of function \code{make_input_range()}}

\item{n_simulations}{Number of rows for output data frame corresponding to a number of simulations.}

\item{discrete}{Logical parameter, if TRUE then random values will be generated from discrete set of values,
if FALSE then random values will be generated from continuous range.}

\item{n_graduations}{Number of discrete values for parameter generation.
Applicable only if discrete is TRUE.}
}
\value{
\code{make_input_format()} returns data frame with a single row corresponding to
a set of current input parameters

\code{make_input_range()} returns a data frame with two rows, the first row is minimal values,
and the second row is maximal values of parameters.

\code{make_input_dataset()} returns data frame with different sets of input parameters
}
\description{
\code{make_input_format()} function allows to prepare a format of dataset of input parameters from
results of a trial simulation. By default it equals all the names of files with some additional arguments: \cr
\code{par_exclude = c(  'censor_cells_number', 'censor_time_step', 'clonefile', 'cloneoutfile', 'ctmax',
'genefile', 'geneoutfile', 'lambda_del', 'lambda_dup', 'logoutfile', 'model_name', 'monitor',
'n_repeat', 'real_time_stop', 'tumbler_for_metastasis_trial', 'tumbler_for_apoptosis_trial',
'tumbler_for_immortalization_trial', 'tumbler_for_angiogenesis_trial', 'tumbler_for_drug_intervention_trial' ) }
}
\section{Functions}{
\itemize{
\item \code{make_input_format()}: Function to prepare a format of dataset of input parameters for parallel calculations

\item \code{make_input_range()}: Function to make the range for each input parameter in the data frame

}}
\examples{
NULL
NULL
NULL
}
