% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_maps.R
\name{make_map}
\alias{make_map}
\title{Function to make a gene_map data.frame with information of genes' locations}
\usage{
make_map(
  f_out = "Input/map.txt",
  ls = c("CCDS4107.1", "CCDS8702.1", "CCDS43171.1", "CCDS11118.1"),
  f_in = "Input/CCDS.current.txt"
)
}
\arguments{
\item{f_out}{Name of file to save gene_map data.frame}

\item{ls}{List of IDs of genes corresponding CCDS database https://ftp.ncbi.nlm.nih.gov/pub/CCDS/current_human/}

\item{f_in}{Name of file to input downloaded from CCDS database}
}
\value{
gene_map data.frame with information of genes' locations for genes of interest
}
\description{
Function to make a gene_map data.frame with information of genes' locations
}
\examples{
url = 'https://ftp.ncbi.nlm.nih.gov/pub/CCDS/current_human/CCDS.current.txt'
download.file( url = url, destfile = 'CCDS.current.txt')
ls   =  c( 'CCDS4107.1', 'CCDS8702.1', 'CCDS43171.1', 'CCDS11118.1' )
gene_map = make_map(f_out    =  'map.txt', ls = ls, f_in =  'CCDS.current.txt' )
}
