% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Mutation_functions.R
\name{mixed_mut_order}
\alias{mixed_mut_order}
\title{Function to get order of mutation for all possible types}
\usage{
mixed_mut_order(clone1)
}
\arguments{
\item{clone1}{Object of class 'Clone'}
}
\value{
data.frame with fields order, type, ID
}
\description{
Function to get order of mutation for all possible types
}
\examples{
clone = tugHall_dataset$clones[[ 46 ]]
clone$PointMut_ID
clone$CNA_ID
pnt_clones = tugHall_dataset$pnt_clones
cna_clones = tugHall_dataset$cna_clones
mixed_mut_order( clone )
}
