% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Mutation_functions.R
\name{modify_gene_map}
\alias{modify_gene_map}
\title{Function to add the mutations to the data.frame gene_map}
\usage{
modify_gene_map(clone1, onco1)
}
\arguments{
\item{clone1}{Object of class 'Clone'}

\item{onco1}{Object of class 'OncoGene'}
}
\value{
list( gm1, gm2 ), where gm1 and gm2 are data.frames gene_maps with mutation information
}
\description{
Function to add the mutations to the data.frame gene_map
}
\examples{
clone = tugHall_dataset$clones[[ 46 ]]
onco = tugHall_dataset$onco_clones[[ 46 ]]
gene_map = tugHall_dataset$gene_map
pnt_clones = tugHall_dataset$pnt_clones
cna_clones = tugHall_dataset$cna_clones
gene_map$pnts = ''
\dontrun{
gm_1_2 = modify_gene_map( clone , onco )  # View(gm_1_2)
}
}
