% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Model_functions.R
\name{number_N_P_M}
\alias{number_N_P_M}
\title{Function to get number of cells of a clone with indicator (normal, primary tumor or metastatic)}
\usage{
number_N_P_M(clone1)
}
\arguments{
\item{clone1}{Object of class 'Clone'}
}
\value{
Vector c( N_normal, N_primary, N_metastatic )
}
\description{
Function to get number of cells of a clone with indicator (normal, primary tumor or metastatic)
}
\examples{
clone1 = tugHall_dataset$clones[[ 1 ]]
number_N_P_M(clone1)
message( paste('Format is as follow: ', 'N_normal', 'N_primary', 'N_metastatic' ) )
}
