% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Model_functions.R
\name{onco_copy}
\alias{onco_copy}
\title{Function to make one copy for onco1 in init_onco_clones function}
\usage{
onco_copy(onco1)
}
\arguments{
\item{onco1}{Object of class 'OncoGene'}
}
\value{
New object of class 'OncoGene' with the same info
}
\description{
Function to make one copy for onco1 in init_onco_clones function
}
\examples{
onco1  =  tugHall_dataset$onco_clones[[ 1 ]]
onco2 = onco_copy( onco1 )  # ID + 1
}
