% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Model_functions.R
\name{onco_update}
\alias{onco_update}
\title{Function to update onco1 after mutation (for usage in trial_mutagenesis() function)}
\usage{
onco_update(onco1, gm)
}
\arguments{
\item{onco1}{Object of class 'OncoGene'}

\item{gm}{data.frame gene_map}
}
\value{
onco1 with updated info
}
\description{
Function to update onco1 after mutation (for usage in trial_mutagenesis() function)
}
\examples{
onco1 = tugHall_dataset$onco_clones[[ 1 ]]
copy_files_to_Input()
define_gene_location()
load_tugHall.Environment( tugHall_dataset )
onco_update( onco1, gm = list(gene_map, gene_map[1:42, ] ) )  # Check CDS length for TP53 gene
}
