% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Classes.R
\docType{class}
\name{OncoGene-class}
\alias{OncoGene-class}
\alias{oncogene}
\title{Class 'OncoGene'}
\description{
Class 'OncoGene'
}
\section{Fields}{

\describe{
\item{\code{id}}{numeric. ID is same as in clone (key for clones)}

\item{\code{name}}{character. Onco genes' names list}

\item{\code{onsp}}{character. Oncogene/suppressor indicator for each gene in list of names}

\item{\code{len}}{numeric. Lengths of onco genes}

\item{\code{cds_1}}{numeric. Onco genes' CDS base lengths for parental chr 1}

\item{\code{cds_2}}{numeric. Onco genes' CDS base lengths for parental chr 2}

\item{\code{rna_1}}{numeric. Onco genes RNA base number length for parental chr 1 (exons+introns)}

\item{\code{rna_2}}{numeric. Onco genes RNA base number length for parental chr 2 (exons+introns)}

\item{\code{p0_1}}{numeric. Probability of absent of mutations for parental chr 1}

\item{\code{p0_2}}{numeric. Probability of absent of mutations for parental chr 2}

\item{\code{prob_1}}{numeric. Vector of relative probabilities for point mutation, deletion and duplication: \cr
prob = c( m0 x sumCDS,  m_del x sumRNA,  m_dup x sumRNA ) / sum( m0 x sumCDS,  m_del x sumRNA,  m_dup x sumRNA )}

\item{\code{prob_2}}{numeric.}

\item{\code{sum_prob_1}}{numeric.}

\item{\code{sum_prob_2}}{numeric.}
}}


\examples{
onco = tugHall_dataset$onco
onco$copy()
}
