% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_maps.R
\name{order_gene_map}
\alias{order_gene_map}
\title{Function to order info in gene_map data.frame with information of genes' locations}
\usage{
order_gene_map(gene_map)
}
\arguments{
\item{gene_map}{data.frame with information of genes' locations}
}
\value{
The same data.frame gene_map with ordered positions for each gene and each chromosome
}
\description{
Function to order info in gene_map data.frame with information of genes' locations
}
\examples{
gene_map = tugHall_dataset$gene_map
gene_map = order_gene_map( gene_map )
}
