% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{pck.env}
\alias{pck.env}
\alias{get_tugHall.Environment}
\alias{load_tugHall.Environment}
\alias{clear_tugHall.Environment}
\title{Environment of the package 'tugHall.3' to store all the objects of a simulation}
\format{
An object of class \code{environment} of length 0.
}
\usage{
pck.env

get_tugHall.Environment()

load_tugHall.Environment(results)

clear_tugHall.Environment()
}
\value{
\code{get_tugHall.Environment} returns all the objects in the pck.env
or tugHall.Environment environment

\code{load_tugHall.Environment} returns NULL and loads
results of simulation to the environment pck.env or tugHall.Environment

\code{clear_tugHall.Environment} returns NULL and clears
the environment pck.env or tugHall.Environment
}
\description{
\code{pck.env } is environment of the package 'tugHall.3'
where all the objects of a simulation are stored and used

\code{get_tugHall.Environment} function returns all the objects
in the pck.env environment of the package tugHall.3

\code{load_tugHall.Environment} loads list
'results' that is results of simulation to the environment pck.env or tugHall.Environment

\code{clear_tugHall.Environment} clears
the environment pck.env or tugHall.Environment
}
\section{Functions}{
\itemize{
\item \code{get_tugHall.Environment}: Get results of simulation stored in pck.env or tugHall.Environment environment

\item \code{load_tugHall.Environment}: Load previous results of simulation to the environment pck.env or tugHall.Environment

\item \code{clear_tugHall.Environment}: Remove all the objects from the environment pck.env or tugHall.Environment
}}

\examples{
NULL
NULL
NULL
}
\keyword{datasets}
