% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_plots.R
\name{plot_2D}
\alias{plot_2D}
\title{Function to plot 2D figure of points y = y(x)}
\usage{
plot_2D(
  x,
  y,
  names = c("X", "Y"),
  pch = 18,
  col = "blue",
  cex = 1.2,
  xr = c(-10, 10),
  yr = c(-10, 10),
  safe_pdf = FALSE,
  filename = "./plot.pdf",
  change_par = TRUE
)
}
\arguments{
\item{x}{Input data for axes X}

\item{y}{Input data for axes Y}

\item{names}{Vector of two characters with names for X and Y axes}

\item{pch}{Parameter pch for plot function}

\item{col}{Colors of points}

\item{cex}{Parameter cex for plot function}

\item{xr}{Range for X}

\item{yr}{Range for Y}

\item{safe_pdf}{Indicator to save plot to a file or not}

\item{filename}{Name of file to save plot if safe_pdf == TRUE}

\item{change_par}{Indicator to change par() or not for a plot. By default change_par = TRUE, after plot it will be returned to initial values}
}
\value{
NULL, making 2D plot using points
}
\description{
Function to plot 2D figure of points y = y(x)
}
\examples{
plot_2D( x=-5:5, y=-3:7 )
}
