% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_plots.R
\name{plot_2D_lines}
\alias{plot_2D_lines}
\title{Function to plot 2D figure of lines  \code{ y_i = DF[, nl[ i ] ] ) }, i - index}
\usage{
plot_2D_lines(
  x,
  DF,
  nl = 1:2,
  names = c("X", "Y"),
  legend_names = "",
  col = c("blue3", "darkmagenta", "red", "green4", "darkorange", "steelblue1"),
  cex = 1.2,
  lwd = 2,
  lt = c(1:6),
  xr = c(-10, 10),
  yr = c(-10, 10),
  safe_pdf = FALSE,
  filename = "./plot.pdf",
  type = "l",
  logscale = "",
  draw_key = TRUE,
  change_par = TRUE
)
}
\arguments{
\item{x}{Input data for axes X}

\item{DF}{data.frame with data to plot}

\item{nl}{indexes of columns in DF to plot}

\item{names}{Vector of two characters with names for X and Y axes}

\item{legend_names}{Name of legend}

\item{col}{Vector of colors for lines}

\item{cex}{Parameter cex for plot function}

\item{lwd}{Vector of width of lines}

\item{lt}{Vector of types of lines}

\item{xr}{Range for X}

\item{yr}{Range for Y}

\item{safe_pdf}{Indicator to save plot to a file or not}

\item{filename}{Name of file to save plot if safe_pdf == TRUE}

\item{type}{Parameter type in plot function}

\item{logscale}{Parameter logscale in plot function}

\item{draw_key}{Indicator to draw key or not}

\item{change_par}{Indicator to change par() or not for a plot. By default change_par = TRUE, after plot it will be returned to initial values}
}
\value{
NULL, making 2D plot using lines
}
\description{
Function to plot 2D figure of lines  \code{ y_i = DF[, nl[ i ] ] ) }, i - index
}
\examples{
DF = tugHall_dataset$data_avg
plot_2D_lines( x = DF[, 1 ], DF, nl = 8:12 , xr = c(1,max(DF$Time) ), yr = c(0,1) )
xr = c(1,max(DF$Time) )
yr = c(0,max(DF[,14],DF[,16],DF[,17] ))
plot_2D_lines( x = DF[, 1 ], DF, nl = c(14,16,17) , xr =xr, yr = yr )
plot_2D_lines( x = DF[, 1 ], DF, nl = 18:22 , xr = c(1,max(DF$Time) ), yr = c(0,1) )
}
