% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_clones.R
\name{plot_average_simulation_data}
\alias{plot_average_simulation_data}
\title{Function to plot main data from data.frame with average data}
\usage{
plot_average_simulation_data(data_avg, time_max)
}
\arguments{
\item{data_avg}{data.frame with average values from cloneout.txt file}

\item{time_max}{Maximal time step in a simulation}
}
\value{
NULL, draw many plot with average data
}
\description{
Function to plot main data from data.frame with average data
}
\examples{
data_avg = tugHall_dataset$data_avg
time_max = tugHall_dataset$time_max
plot_average_simulation_data( data_avg , time_max = time_max )
}
