% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_plots.R
\name{plot_clone_evolution}
\alias{plot_clone_evolution}
\title{Function to plot clone evolution}
\usage{
plot_clone_evolution(
  data_flow,
  threshold = c(0.05, 1),
  lwd = 2,
  hue = c(" ", "random", "red", "orange", "yellow", "green", "blue", "purple", "pink",
    "monochrome")[1],
  luminosity = c(" ", "random", "light", "bright", "dark")[5],
  yr = NA,
  add_initial = TRUE,
  log_scale = FALSE,
  change_par = TRUE
)
}
\arguments{
\item{data_flow}{data.frame with results of simulation at each time step}

\item{threshold}{Vector two numbers from 0 to 1 to show clones with relative final numbers of cells in the range of threshold}

\item{lwd}{Line width in the plot function}

\item{hue}{Parameter hue in the function randomColor from library randomcoloR}

\item{luminosity}{Parameter luminosity in the function randomColor from library randomcoloR}

\item{yr}{Range for Y axes}

\item{add_initial}{Indicator to add or do not add initial clones to plot}

\item{log_scale}{Indicator to use log_scale or not for Y axes}

\item{change_par}{Indicator to change par() or not for a plot. By default change_par = TRUE, after plot it will be returned to initial values}
}
\value{
NULL, making plot with clones evolution
}
\description{
Function to plot clone evolution
}
\examples{
data_flow = tugHall_dataset$data_flow
plot_clone_evolution( data_flow, threshold = c(0.01, 1 ), add_initial = TRUE, log_scale = FALSE )
plot_clone_evolution( data_flow, threshold = c(0, 0.01 ), add_initial = FALSE, log_scale = TRUE )
}
