% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_plots.R
\name{plot_order_dysfunction}
\alias{plot_order_dysfunction}
\title{Function to plot order of genes dysfunction as a step function with number of cells related to each order}
\usage{
plot_order_dysfunction(
  rdr_dysf,
  pos = c(0, 100),
  logscale = "y",
  cex = 1,
  change_par = TRUE
)
}
\arguments{
\item{rdr_dysf}{Order of genes dysfunction as a data.frame}

\item{pos}{Coordinates of list of order of genes dysfunction}

\item{logscale}{Parameter logscale for plot function}

\item{cex}{Parameter cex for plot function}

\item{change_par}{Indicator to change par() or not for a plot. By default change_par = TRUE, after plot it will be returned to initial values}
}
\value{
NULL, making plot with step function of order of genes' dysfunction
}
\description{
Function to plot order of genes dysfunction as a step function with number of cells related to each order
}
\examples{
rdr_dysf = tugHall_dataset$rdr_dysf
plot_order_dysfunction( rdr_dysf , logscale = '', pos = c(8, 5000), cex = 1.4)
plot_order_dysfunction( rdr_dysf , logscale = 'y', pos = c(10, 100), cex = 1.2)
}
