% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_plots.R
\name{plot_VAF}
\alias{plot_VAF}
\title{Function to plot the distributions of VAF for each gene after simulation}
\usage{
plot_VAF(
  VAF,
  rho = 0,
  violin = FALSE,
  save_to_file = FALSE,
  file_name = "./plot_VAF.pdf",
  wait_for_user = FALSE
)
}
\arguments{
\item{VAF}{is Variant allele frequencies of genes in the output format of the function \code{get_rho_VAF}}

\item{rho}{is \code{rho} value of VAF.}

\item{violin}{Logical parameter to draw the distribution in the form of violin or box plot.
By default \code{violin = FALSE}, i.e. it draws in the form of box plot.}

\item{save_to_file}{Logical parameter to save or do not save plot to the file. by default \code{save_to_file = FALSE}}

\item{file_name}{Name of file to save plot. By default \code{file_name = './plot_VAF.pdf' }}

\item{wait_for_user}{Logical parameter to stop at each plot or do not stop. By default \code{wait_for_user = FALSE }}
}
\value{
\code{plot_VAF()} function returns NULL making plot with VAF distributions for each gene
}
\description{
\code{plot_VAF()} function draw the distributions of VAF for each gene after simulation
}
\examples{

NULL
}
