% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Mutation_functions.R
\name{pnts_add_dlt}
\alias{pnts_add_dlt}
\title{Function to subtract delta from position of point mutations}
\usage{
pnts_add_dlt(gm_w1, dlt)
}
\arguments{
\item{gm_w1}{A row from data.frame gene_map}

\item{dlt}{Delta to subtract from positions of point mutations}
}
\value{
Return the pnts - dlt for one row of data.frame gene_map
}
\description{
Function to subtract delta from position of point mutations
}
\examples{
gene_map = tugHall_dataset$gene_map
gene_map$pnts = ''
gene_map$pnts[6] = '112792451'
gm_w1 = gene_map[6,]
pnts_add_dlt( gm_w1 , dlt = 1000 )
pnts_add_dlt( gm_w1 , dlt = -1001 )
}
