% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Classes.R
\docType{class}
\name{Point_Mutations-class}
\alias{Point_Mutations-class}
\alias{Point_Mutations}
\title{Class 'Point_Mutations'}
\description{
Class 'Point_Mutations'
}
\section{Fields}{

\describe{
\item{\code{PointMut_ID}}{numeric. ID of point mutation}

\item{\code{Allele}}{character. A or B allele}

\item{\code{Parental_1or2}}{numeric. Parental chromosome, could be 1 or 2}

\item{\code{Chr}}{character. Chromosome name}

\item{\code{Ref_pos}}{numeric. Reference position}

\item{\code{Phys_pos}}{vector. Physical positions}

\item{\code{Delta}}{vector. Delta of positions}

\item{\code{Copy_number}}{numeric. Copy number of allele}

\item{\code{Gene_name}}{character. Gene's name}

\item{\code{MalfunctionedByPointMut}}{logical. True for driver mutation and False for passenger mutation}

\item{\code{mut_order}}{numeric. Number in order of mutation to reproduce the gene_map data.frame}
}}


\examples{
pnt = tugHall_dataset$pnt_clones[[ 1 ]]
print( pnt )
pnt$copy()
pnt$show()
pnt$initialize()
pnt$show()
pnt = tugHall_dataset$pnt_clones[[ 3 ]]
pnt$safe()   # save as row of data.frame
}
