% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Input.R
\name{print_parameters}
\alias{print_parameters}
\title{Function to print GLOBAL parameters}
\usage{
print_parameters()
}
\value{
Message with values of all the GLOBAL parameters
}
\description{
Function to print GLOBAL parameters
}
\examples{
copy_files_to_Input()
define_parameters( read_fl = FALSE )
define_compaction_factor()
print_parameters()
}
