% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_file}
\alias{read_file}
\title{Function to read file}
\usage{
read_file(file_name = "", stringsAsFactors = FALSE, header = TRUE)
}
\arguments{
\item{file_name}{Name of file to read}

\item{stringsAsFactors}{Parameter for read.table function, by default stringsAsFactors = FALSE}

\item{header}{Logical type to read or do not read head of a file}
}
\value{
data.frame of data from a file
}
\description{
Function to read file
}
\examples{
fl = system.file('extdata/Input', 'gene_map.txt',package = 'tugHall.3', mustWork = TRUE )
read_file(file_name = fl, stringsAsFactors = FALSE )
fl = system.file('extdata/Input', 'CF.txt',package = 'tugHall.3', mustWork = TRUE )
read_file(file_name = fl, stringsAsFactors = FALSE, header = FALSE )
}
