% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Mutation_functions.R
\name{safe_pnt_mut}
\alias{safe_pnt_mut}
\title{Function to save 1 point mutation in a data frame}
\usage{
safe_pnt_mut(pnt)
}
\arguments{
\item{pnt}{Object of class 'Point_Mutations'}
}
\value{
data frame with 1 row of point mutation info
}
\description{
Function to save 1 point mutation in a data frame
}
\examples{
pnt = tugHall_dataset$pnt_clones[[ 1 ]]
df = safe_pnt_mut( pnt ) # View( pnt )
}
