% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{simulation}
\alias{simulation}
\alias{simulation_keep_run}
\title{Simulation for lazy start with parameters from Input folder}
\usage{
simulation(
  verbose = TRUE,
  to_plot = TRUE,
  seed = NA,
  work_dir = getwd(),
  digits = 6
)

simulation_keep_run(
  loadRDS = TRUE,
  fileRDS = "./Results_of_simulation.RDS",
  change_parameters = list(),
  seed = NA,
  work_dir = getwd(),
  digits = 6
)
}
\arguments{
\item{verbose}{Logical type to show or do not show messages during execution}

\item{to_plot}{Logical type to plot or do not plot graphical results of a simulation}

\item{seed}{Numeric type to set seed for a simulation, if seed = NA (by default) then it will be skipped}

\item{work_dir}{Working directory for a simulation, by default \code{ work_dir = getwd() }}

\item{digits}{Number of digits in the numeric format of pck.env environment of tugHall}

\item{loadRDS}{logical to load data of previous simulation from file fileRDS.
If \code{loadRDS = FALSE} then it loads data from pck.env that should contain the data of a simulation.}

\item{fileRDS}{file name to load data of previous simulation, only if \code{loadRDS = TRUE}}

\item{change_parameters}{List of parameters to change from the previous simulation,
each parameter should be corresponding to variable name.
For example, \code{change_parameters = list(censore_n = 1E06, censore_t = 60 ) }}
}
\value{
List of results of simulation with default values for all the parameters

List of (clones, onco_clones), where clones - list of objects of class 'Clone', and onco_clones - list of objects of class 'OncoGene'. During a simulation it saves data to geneoutfile.
}
\description{
\code{simulation()} makes a simulation with parameters from Input folder
and results save in pck.env as well in './Results_of_simulation.RDS' file in \code{work_dir} folder

\code{simulation_keep_run()} is needed to start simulation from previous results with new parameter set.
Parameter set can be defined as usually from Input folder or keep all the parameters excluding
input list of parameters.
}
\section{Functions}{
\itemize{
\item \code{simulation_keep_run}: \code{simulation_keep_run()} is needed to start simulation from previous results with new parameter set
}}

\examples{

# it takes a time for a simulation and then it will demonstrates results, \cr
# so, please, wait for a while
simulation( verbose = FALSE , to_plot = FALSE )
NULL
\dontrun{


}
}
