% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{simulation_example}
\alias{simulation_example}
\title{Example of simulation for lazy start}
\usage{
simulation_example(
  verbose = TRUE,
  to_plot = TRUE,
  seed = NA,
  work_dir = getwd(),
  digits = 5
)
}
\arguments{
\item{verbose}{Logical type to show or do not show messages during execution}

\item{to_plot}{Logical type to plot or do not plot graphical results of a simulation}

\item{seed}{Numeric type to set seed for a simulation, if seed = NA (by default) then it will be skipped}

\item{work_dir}{Working directory for a simulation, by default \code{ work_dir = getwd() }}

\item{digits}{Number of digits in the numeric format of pck.env environment of tugHall}
}
\value{
List of results of simulation with default values for all the parameters
}
\description{
Example of simulation for lazy start
}
\examples{

# it takes a time for a simulation and then it will demonstrates results, \cr
# so, please, wait for a while
simulation_example( verbose = FALSE , to_plot = FALSE )
}
