% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Model_functions.R
\name{sum_cell}
\alias{sum_cell}
\title{Aggregate data of a clone for environment object}
\usage{
sum_cell(env, clones)
}
\arguments{
\item{env}{Object of class 'Environ'}

\item{clones}{List of all the objects of class 'Clone'}
}
\value{
NULL, but global variable env is updated
}
\description{
Aggregate data of a clone for environment object
}
\examples{
clones = tugHall_dataset$clones
env = tugHall_dataset$env
sum_cell(env, clones)
message( paste0('Number of primary tumor cells in the pool of clones is ', env$P ) )
message( paste0('Number of normal cells in the pool of clones is ', env$N ) )
message( paste0('Number of metastatic cells in the pool of clones is ', env$M ) )
}
