% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Model_functions.R
\name{sum_mutation}
\alias{sum_mutation}
\title{Serve function for sum_cell() function}
\usage{
sum_mutation(clone1)
}
\arguments{
\item{clone1}{Object of class 'Clone'}
}
\value{
vector of clone1 variables to aggregate in sum_cell() function
}
\description{
Serve function for sum_cell() function
}
\examples{
clone1 = tugHall_dataset$clones[[ 1 ]]
sum_mutation(clone1)
}
