% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Model_functions.R
\name{sum_N_P_M}
\alias{sum_N_P_M}
\title{Function to calculate N and M numbers - normal and metastatic cells}
\usage{
sum_N_P_M(env, clones)
}
\arguments{
\item{env}{Object of class 'Environ'}

\item{clones}{List of all the objects of class 'Clone'}
}
\value{
Number of all the cells in a simulation (normal + primary tumor + metastatic)
}
\description{
Function to calculate N and M numbers - normal and metastatic cells
}
\examples{
clones = tugHall_dataset$clones
env = tugHall_dataset$env
env$M = 0
env$P = 0
env$N = 0  # View( env )
sum_N_P_M(env, clones)  # View( env )
message( paste(env$N, env$P, env$M ) )
}
