% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Model_functions.R
\name{trial_mutagenesis}
\alias{trial_mutagenesis}
\title{Function for mutagenesis trial}
\usage{
trial_mutagenesis(clone1, num_mut, onco1)
}
\arguments{
\item{clone1}{Object of class 'Clone'}

\item{num_mut}{Number of mutations in this NEW clone1}

\item{onco1}{Object of class 'OncoGene' corresponding to clone1 (with the same ID)}
}
\value{
Changed object clone1, add related mutations to the lists of point mutations and/or CNA mutations
}
\description{
Function for mutagenesis trial
}
\examples{
copy_files_to_Input()
copy_files_to_Output()
define_parameters()
clone1 = tugHall_dataset$clones[[ 1 ]]
onco1 = tugHall_dataset$onco_clones[[ 1 ]]
onco = tugHall_dataset$onco
define_gene_location()
pnt_clones = tugHall_dataset$pnt_clones
cna_clones = tugHall_dataset$cna_clones
mut_order = 234  # Just an example number
message( c('CNA mutation IDs ', paste(clone1$CNA_ID, collapse = ' ') ) )
message( c('Point mutation IDs ', paste(clone1$PointMut_ID, collapse = ' ') ) )
\dontrun{
trial_mutagenesis( clone1, num_mut = 1, onco1 )  # it adds info to clone1
message( c('CNA mutation IDs ', paste(clone1$CNA_ID, collapse = ' ') ) )
message( c('Point mutation IDs ', paste(clone1$PointMut_ID, collapse = ' ') ) )

trial_mutagenesis( clone1, num_mut = 10, onco1 )  # it adds info to clone1
message( c('CNA mutation IDs ', paste(clone1$CNA_ID, collapse = ' ') ) )
message( c('Point mutation IDs ', paste(clone1$PointMut_ID, collapse = ' ') ) )
}
}
