% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall.3-package.R
\docType{package}
\name{tugHall.3-package}
\alias{tugHall.3}
\alias{tugHall.3-package}
\title{tugHall.3: R-based script to simulate the cancer cell evolution}
\description{
tugHall (tumor gene-Hallmark) is a cancer-cell evolution model simulator, wherein gene mutations are linked to the hallmarks of cancer, which influence tumor cell behaviors.
}
\author{
\strong{Maintainer}: Iurii Nagornov \email{inagonov@ncc.go.jp}

Authors:
\itemize{
  \item Mamoru Kato \email{mamkato@ncc.go.jp}
}

Other contributors:
\itemize{
  \item Jo Nishino \email{jnishino@ncc.go.jp} [contributor]
  \item Eisaku Furukawa \email{efurukaw@ncc.go.jp} [contributor]
}

}
\keyword{internal}
