% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Model_functions.R
\name{write_cloneout}
\alias{write_cloneout}
\title{Function to write data to cloneout file at a time step}
\usage{
write_cloneout(outfile, env, clones, isFirst, onco_clones)
}
\arguments{
\item{outfile}{File name for output info}

\item{env}{Object of class 'Environ'}

\item{clones}{List of objects of class 'Clone'}

\item{isFirst}{logical type = TRUE as default}

\item{onco_clones}{List of objects of class 'OncoGene'}
}
\value{
NULL, but add rows to output file with clone evolution data
}
\description{
Function to write data to cloneout file at a time step
}
\examples{
env = tugHall_dataset$env
onco = tugHall_dataset$onco
if ( !dir.exists('./Output') ) dir.create('./Output')
clones = tugHall_dataset$clones
onco_clones = tugHall_dataset$onco_clones
write_header(outfile='./Output/exmpl.txt', env, onco)
write_cloneout( outfile = './Output/exmpl.txt', env, clones, isFirst = TRUE, onco_clones )
}
