% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Model_functions.R
\name{write_geneout}
\alias{write_geneout}
\title{Function to write info about HallMark data}
\usage{
write_geneout(outfile, hall, Compaction_factor, CF)
}
\arguments{
\item{outfile}{File name for output info}

\item{hall}{Object of class "HallMark"}

\item{Compaction_factor}{Compaction factor, logical type only. True means 'to use', False means 'do not use' Compaction factor for hallmarks variables}

\item{CF}{Vector with values of compaction factor for each hallmark}
}
\value{
NULL, but data will save to a file
}
\description{
Function to write info about HallMark data
}
\examples{
copy_files_to_Input()
load_tugHall.Environment( tugHall_dataset )
if ( !dir.exists('./Output') ) dir.create('./Output')
withr::with_environment( env = pck.env, code =  write_geneout(outfile = geneoutfile, hall, Compaction_factor, CF) )
}
