% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Model_functions.R
\name{write_header}
\alias{write_header}
\title{Function to write the header to a file}
\usage{
write_header(outfile, env, onco)
}
\arguments{
\item{outfile}{File name for output info}

\item{env}{Object of class 'Environ'}

\item{onco}{Object of class "OncoGene"}
}
\value{
NULL, but the header will save to a file and delete old info
}
\description{
Function to write the header to a file
}
\examples{
env = tugHall_dataset$env
onco = tugHall_dataset$onco
if ( !dir.exists('./Output') ) dir.create('./Output')
write_header(outfile='./Output/exmpl.txt', env, onco)
}
