% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Model_functions.R
\name{write_log}
\alias{write_log}
\title{Function to write log file}
\usage{
write_log(
  genefile,
  clonefile,
  geneoutfile,
  cloneoutfile,
  logoutfile,
  E0,
  F0,
  m0,
  uo,
  us,
  s0,
  k0,
  m_dup,
  m_del,
  lambda_dup,
  lambda_del,
  uo_dup,
  us_dup,
  uo_del,
  us_del,
  censore_n,
  censore_t,
  d0,
  Compaction_factor,
  model_name,
  time_stop,
  n_repeat,
  monitor
)
}
\arguments{
\item{genefile}{File name of initial OncoGene information}

\item{clonefile}{File name of info about initial clones}

\item{geneoutfile}{File name for output info about OncoGene information}

\item{cloneoutfile}{File name for output info with clone evolution data}

\item{logoutfile}{Name of log file with all the parameters}

\item{E0}{Parameter in the division probability, numeric type only}

\item{F0}{Parameter in the division probability, numeric type only}

\item{m0}{Mutation probability for point mutation, numeric type only}

\item{uo}{Oncogene mutation probability, numeric type only}

\item{us}{Suppressor mutation probability, numeric type only}

\item{s0}{Parameter in the sigmoid function, numeric type only}

\item{k0}{Environmental death probability, numeric type only}

\item{m_dup}{Mutation probability for duplication, numeric type only}

\item{m_del}{Mutation probability for deletion, numeric type only}

\item{lambda_dup}{CNA duplication average length (of the geometrical distribution for the length), integer type only}

\item{lambda_del}{CNA deletion average length (of the geometrical distribution for the length), integer type only}

\item{uo_dup}{Gene malfunction probability by CNA duplication for oncogene, numeric type only}

\item{us_dup}{Gene malfunction probability by CNA duplication for suppressor, numeric type only}

\item{uo_del}{Gene malfunction probability by CNA deletion    for oncogene, numeric type only}

\item{us_del}{Gene malfunction probability by CNA deletion    for suppressor, numeric type only}

\item{censore_n}{Max cell number where the program forcibly stops, integer type only}

\item{censore_t}{Max time where the program forcibly stops, integer type only}

\item{d0}{Initial probability to divide cells, numeric type only}

\item{Compaction_factor}{Compaction factor, logical type only. True means 'to use', False means 'do not use' Compaction factor for hallmarks variables}

\item{model_name}{Name of the model to use. Can be  'proportional_metastatic' or 'threshold_metastatic' or 'simplified'}

\item{time_stop}{Max time in seconds of running after that the program forcibly stops, integer type only}

\item{n_repeat}{Max number of repetition of the program until the NON-ZERO output will be getting, integer type only}

\item{monitor}{The indicator to make monitor file during a simulation or do not make, logical type only}
}
\value{
NULL, write log file to Output folder
}
\description{
Function to write log file
}
\examples{
copy_files_to_Input()
define_files_names()
define_parameters()
if ( !dir.exists('./Output') ) dir.create('./Output')
\dontrun{
write_log(genefile, clonefile, geneoutfile, cloneoutfile, logoutfile,
E0, F0, m0, uo, us, s0, k0, m_dup, m_del, lambda_dup, lambda_del,
uo_dup, us_dup, uo_del, us_del, censore_n, censore_t, d0,
Compaction_factor, model_name, time_stop, n_repeat, monitor )
}
}
