% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Model_functions.R
\name{write_pnt_clones}
\alias{write_pnt_clones}
\title{Function to write the point mutation info for all clones for all time steps, used at the last time step or after simulation}
\usage{
write_pnt_clones(pnt_clones, file_out = "Output/point_mutations.txt")
}
\arguments{
\item{pnt_clones}{List of objects of class 'Point_Mutations'}

\item{file_out}{File name to write}
}
\value{
NULL, but info will write to a file
}
\description{
Function to write the point mutation info for all clones for all time steps, used at the last time step or after simulation
}
\examples{
pnt_clones = tugHall_dataset$pnt_clones
if ( !dir.exists('./Output') ) dir.create('./Output')
write_pnt_clones(pnt_clones, file_out = 'Output/point_mutations.txt')
}
