% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Model_functions.R
\name{clone_copy}
\alias{clone_copy}
\title{Function to make one copy for clone1 in clone_init function}
\usage{
clone_copy(clone1)
}
\arguments{
\item{clone1}{Object of class 'Clone'}
}
\value{
New object of class 'Clone' with the same info and new ID
}
\description{
Function to make one copy for clone1 in clone_init function
}
\examples{
clone1 = tugHall_dataset$clones[[ 1 ]]
env = tugHall_dataset$env
define_parameters()
clone_copy(clone1)
}
