% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_clones.R
\name{get_order_of_genes_dysfunction}
\alias{get_order_of_genes_dysfunction}
\title{Function to get order of genes' dysfunction}
\usage{
get_order_of_genes_dysfunction(
  pnt_mut,
  data_last,
  cna_mut,
  file_name = "./Output/order_genes_dysfunction.txt"
)
}
\arguments{
\item{pnt_mut}{data.frame with info about all the point mutations}

\item{data_last}{data.frame with data of simulation at the last time step}

\item{cna_mut}{data.frame with info about all the CNA mutations}

\item{file_name}{Name of file to save data}
}
\value{
data.frame of genes' dysfunction and save it in a file
}
\description{
Function to get order of genes' dysfunction
}
\examples{
pnt_mut  =  tugHall_dataset$pnt_mut
data_last  =  tugHall_dataset$data_last
cna_mut = tugHall_dataset$cna_mut
file_name = './Output/order_genes_dysfunction.txt'
rdr = get_order_of_genes_dysfunction( pnt_mut, data_last, cna_mut, file_name = file_name )
}
