% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tugHall.Environment}
\alias{get_tugHall.Environment}
\title{Get results of simulation storaged in pck.env or tugHall.Environment environment}
\usage{
get_tugHall.Environment()
}
\description{
Get results of simulation storaged in pck.env or tugHall.Environment environment
}
\examples{
NULL
}
