% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Mutation_functions.R
\name{get_u_cna}
\alias{get_u_cna}
\title{Function to choose probability of CNA mutation for several genes}
\usage{
get_u_cna(genes, dupOrdel)
}
\arguments{
\item{genes}{Names of genes, vector of names}

\item{dupOrdel}{It could be 'dup' or 'del' to denote duplication or deletion}
}
\value{
Single value of maximal probability from probabilities for several genes
}
\description{
Function to choose probability of CNA mutation for several genes
}
\examples{
copy_files_to_Input()
load_tugHall.Environment( tugHall_dataset )
onco = tugHall_dataset$onco
withr::with_environment( env = pck.env, code = get_u_cna( genes = 'APC', dupOrdel = 'dup' ) )
withr::with_environment( env = pck.env, code = get_u_cna( genes = c('KRAS','APC'), dupOrdel = c('dup', 'del') ) )
}
