% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Model_functions.R
\name{init_clones}
\alias{init_clones}
\title{Function to read file with initial clones}
\usage{
init_clones(clonefile, clone1)
}
\arguments{
\item{clonefile}{File to read}

\item{clone1}{Object of class 'Clone'}
}
\value{
List of objects of class 'Clone
}
\description{
Function to read file with initial clones
}
\examples{
copy_files_to_Input()
load_tugHall.Environment( tugHall_dataset )
clone1 = tugHall_dataset$clones[[ 1 ]]
withr::with_environment( env = pck.env, code =  clones = init_clones(clonefile, clone1) )
}
