% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Model_functions.R
\name{init_onco_clones}
\alias{init_onco_clones}
\title{Function to make list of objects of class 'OncoGene' and generate initial onco settings for all clones (onco_clones)}
\usage{
init_onco_clones(onco1, clones)
}
\arguments{
\item{onco1}{Object of class 'OncoGene'}

\item{clones}{List of objects of class 'Clone'}
}
\value{
List of objects of class 'OncoGene'
}
\description{
Function to make list of objects of class 'OncoGene' and generate initial onco settings for all clones (onco_clones)
}
\examples{
copy_files_to_Input()
define_files_names()
env = tugHall_dataset$env
load_tugHall.Environment( tugHall_dataset )
onco = tugHall_dataset$onco
clone1 = tugHall_dataset$clones[[ 1 ]]
clones = init_clones(clonefile, clone1)  # View( clones )
onco_clones = init_onco_clones( onco1 = onco, clones )
}
