% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{make_input_dataset}
\alias{make_input_dataset}
\title{Function to prepare dataset of input parameters for parallel calculations}
\usage{
make_input_dataset()
}
\value{
\code{make_input_dataset()} returns data frame with different sets of input parameters
}
\description{
Function to prepare dataset of input parameters for parallel calculations
}
\examples{
NULL
}
