% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{make_input_format}
\alias{make_input_format}
\title{Function to prepare a format of dataset of input parameters for parallel calculations}
\usage{
make_input_format(par_exclude = list())
}
\arguments{
\item{par_exclude}{List of parameters to exclude from data frame of
input parameters because they will be constant for all the simulations}
}
\value{
\code{make_input_format()} returns data frame with a single row corresponding to
a set of current input parameters
}
\description{
\code{make_input_format()} function allows to prepare a format of dataset of input parameters from
results of a trial simulation. By default it is empty.
}
\examples{
NULL
}
