% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_plots.R
\name{plot_2D}
\alias{plot_2D}
\alias{plot_2D_lines}
\alias{plot_order_dysfunction}
\alias{plot_clone_evolution}
\title{Function to plot 2D figure of lines}
\usage{
plot_2D(
  x,
  y,
  names = c("X", "Y"),
  pch = 18,
  col = "blue",
  cex = 1.2,
  xr = c(-10, 10),
  yr = c(-10, 10),
  safe_pdf = FALSE,
  filename = "./plot.pdf",
  par_list = list(xpd = TRUE, cex.lab = 2, lwd = 2, mar = c(5, 5, 5, 5), tcl = 0.5,
    cex.axis = 1.75, mgp = c(2.2, 0.5, 0), font.axis = 2, font.lab = 2)
)

plot_2D_lines(
  x,
  DF,
  nl = 1:2,
  names = c("X", "Y"),
  legend_names = "",
  col = c("blue3", "darkmagenta", "red", "green4", "darkorange", "steelblue1"),
  cex = 1.2,
  lwd = 2,
  lt = c(1:6),
  xr = c(-10, 10),
  yr = c(-10, 10),
  safe_pdf = FALSE,
  filename = "./plot.pdf",
  type = "l",
  logscale = "",
  draw_key = TRUE,
  par_list = list(xpd = TRUE, cex.lab = 2, lwd = 2, mar = c(5, 5, 5, 5), tcl = 0.5,
    cex.axis = 1.75, mgp = c(2.2, 0.5, 0), font.axis = 2, font.lab = 2),
  cex.legend = 1.3
)

plot_order_dysfunction(
  rdr_dysf,
  pos = c(0, 100),
  logscale = "y",
  cex = 1,
  par_list = list(xpd = TRUE, cex.lab = 2, lwd = 2, mar = c(5, 5, 5, 5), tcl = 0.5,
    cex.axis = 1.75, mgp = c(2.2, 0.5, 0), font.axis = 2, font.lab = 2),
  cex.legend = 1.3
)

plot_clone_evolution(
  data_flow,
  threshold = c(0.05, 1),
  lwd = 2,
  hue = c(" ", "random", "red", "orange", "yellow", "green", "blue", "purple", "pink",
    "monochrome")[1],
  luminosity = c(" ", "random", "light", "bright", "dark")[5],
  yr = NA,
  add_initial = TRUE,
  log_scale = FALSE,
  par_list = list(xpd = TRUE, cex.lab = 2, lwd = 2, mar = c(5, 5, 5, 5), tcl = 0.5,
    cex.axis = 1.75, mgp = c(2.2, 0.5, 0), font.axis = 2, font.lab = 2)
)
}
\arguments{
\item{x}{Input data for axes X}

\item{y}{Input data for axes Y}

\item{names}{Vector of two characters with names for X and Y axes}

\item{pch}{Parameter pch for plot function corresponding types of dots}

\item{col}{Vector of colors for lines or dots}

\item{cex}{Parameter cex for plot function}

\item{xr}{Range for X}

\item{yr}{Range for Y}

\item{safe_pdf}{Indicator to save plot to a file or not}

\item{filename}{Name of file to save plot if safe_pdf == TRUE}

\item{par_list}{List of parameters to set locally for \code{par()} function.
By default \code{ par_list = list( xpd=TRUE, cex.lab=2, lwd = 2, mar = c(5, 5, 5, 5), } \cr
\code{ tcl = 0.5, cex.axis = 1.75,  mgp = c(2.2, 0.5, 0), font.axis = 2, font.lab = 2 )}}

\item{DF}{data.frame with data to plot}

\item{nl}{indexes of columns in DF to plot}

\item{legend_names}{Name of legend}

\item{lwd}{Vector of width of lines}

\item{lt}{Vector of types of lines}

\item{type}{Parameter type in plot function}

\item{logscale}{Parameter logscale in plot function, can be '' or 'y' or 'x'}

\item{draw_key}{Indicator to draw key or not}

\item{cex.legend}{Character expansion factor for text of legend on the plot}

\item{rdr_dysf}{Order of genes dysfunction as a data.frame}

\item{pos}{Coordinates of list of order of genes dysfunction}

\item{data_flow}{data.frame with results of simulation at each time step}

\item{threshold}{Vector two numbers from 0 to 1 to show clones with relative final numbers of cells in the range of threshold}

\item{hue}{Parameter hue in the function randomColor from library randomcoloR. \cr
\code{hue = c(" ", "random", "red", "orange", "yellow", "green", "blue", "purple", "pink", "monochrome")[1]},
so by default ' ' (blank space)}

\item{luminosity}{Parameter luminosity in the function randomColor from library randomcoloR. \cr
It can be \code{luminosity = c(" ", "random", "light", "bright", "dark")[5]}, so by default 'dark'}

\item{add_initial}{Logical indicator to add or do not add initial clones to plot}

\item{log_scale}{Logical indicator to use logarithmic scale or not for Y axes}
}
\value{
\code{plot_2D()} function returns NULL, making 2D plot using points

NULL, making 2D plot using lines

\code{plot_order_dysfunction()} returns NULL making plot with step function of order of genes' dysfunction

\code{plot_clone_evolution()} function returns NULL making plot with clones evolution
}
\description{
\code{plot_2D()} function used to plot 2D figure of points y = y(x)

\code{plot_2D_lines()} function returns NULL and plot 2D figure of lines from data.frame DF like \cr
\code{ y_i = DF[, nl[ i ] ] ) }, nl - indexes of columns

\code{plot_order_dysfunction()} function draw the order of genes dysfunction
as a step function with number of cells related to each order

\code{plot_clone_evolution()} function draw the clones' evolution as cells numbers for each clone
}
\section{Functions}{
\itemize{
\item \code{plot_2D}: Function to plot 2D figure of points y = y(x)

\item \code{plot_order_dysfunction}: Function to plot order of genes dysfunction as a step function with number of cells related to each order

\item \code{plot_clone_evolution}: Function to plot clone evolution
}}

\examples{
plot_2D( x=-5:5, y=-3:7 )
DF = tugHall_dataset$data_avg
plot_2D_lines( x = DF[, 1 ], DF, nl = 8:12 , xr = c(1,max(DF$Time) ), yr = c(0,1) )
xr = c(1,max(DF$Time) )
yr = c(0,max(DF[,14],DF[,16],DF[,17] ))
plot_2D_lines( x = DF[, 1 ], DF, nl = c(14,16,17) , xr =xr, yr = yr )
plot_2D_lines( x = DF[, 1 ], DF, nl = 18:22 , xr = c(1,max(DF$Time) ), yr = c(0,1) )
rdr_dysf = tugHall_dataset$rdr_dysf
plot_order_dysfunction( rdr_dysf , logscale = '', pos = c(8, 5000), cex = 1.4)
plot_order_dysfunction( rdr_dysf , logscale = 'y', pos = c(10, 100), cex = 1.2)
data_flow = tugHall_dataset$data_flow
plot_clone_evolution( data_flow, threshold = c(0.01, 1 ), add_initial = TRUE, log_scale = FALSE )
plot_clone_evolution( data_flow, threshold = c(0, 0.01 ), add_initial = FALSE, log_scale = TRUE )
}
