% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tugHall_Model_functions.R
\name{write_weights}
\alias{write_weights}
\title{Function to write info about relationship between genes and hallmarks}
\usage{
write_weights(outfile, hall)
}
\arguments{
\item{outfile}{File name for output info}

\item{hall}{Object of class 'HallMark'}
}
\value{
NULL, but info about relationship between genes and hallmarks will write to a file
}
\description{
Function to write info about relationship between genes and hallmarks
}
\examples{
if ( !dir.exists('./Output') ) dir.create('./Output')
hall = tugHall_dataset$hall
onco = tugHall_dataset$onco
write_weights(outfile = './Output/weights.txt', hall)
}
